move hashes/fuzz into main fuzz/ directory
