+++ fuzz/fuzz_targets/hashes/ripemd160.rs
@@ -1,19 +1,16 @@
 
-extern crate bitcoin_hashes;
-extern crate crypto;
-
-use bitcoin_hashes::Hash;
-use bitcoin_hashes::sha1;
+use bitcoin::hashes::Hash;
+use bitcoin::hashes::ripemd160;
 use crypto::digest::Digest;
-use crypto::sha1::Sha1;
+use crypto::ripemd160::Ripemd160;
 
 fn do_test(data: &[u8]) {
-    let our_hash = sha1::Hash::hash(data);
+    let our_hash = ripemd160::Hash::hash(data);
 
     let mut rc_hash = [0u8; 20];
-    let mut rc_sha1 = Sha1::new();
-    rc_sha1.input(data);
-    rc_sha1.result(&mut rc_hash);
+    let mut rc_engine = Ripemd160::new();
+    rc_engine.input(data);
+    rc_engine.result(&mut rc_hash);
 
     assert_eq!(&our_hash[..], &rc_hash[..]);
 }
